refactor: move documentContext to `@volar/vue-language-service`
