+++ packages/vue-typescript/package.json
@@ -23,7 +23,6 @@
 		"@volar/vue-code-gen": "0.33.2",
 		"@vue/reactivity": "^3.2.27",
 		"upath": "^2.0.1",
-		"vscode-css-languageservice": "^5.1.9",
 		"vscode-html-languageservice": "^4.2.1",
 		"vscode-languageserver-protocol": "^3.17.0-next.12",
 		"vscode-languageserver-textdocument": "^1.0.3"
